Created UserForm component and pass values for sign up or register. Tested and working
